refactor(naming): update filenames to be consistent
